chore: rename webpush pkg
