+++ services/web_notify/subscribe_handler/handler.go
@@ -7,7 +7,7 @@ import (
 
 	mongolib "github.com/eugbyte/monorepo/libs/db/mongo_lib"
 	"github.com/eugbyte/monorepo/libs/formats"
-	"github.com/eugbyte/monorepo/services/web-push/models"
+	"github.com/eugbyte/monorepo/services/webnotify/models"
 	"go.mongodb.org/mongo-driver/bson"
 )
